final demo app

+++ app/src/main/java/com/alperen/cryptocurrencydemoapp/model/ticker/Volumes.kt
@@ -4,7 +4,7 @@ package com.alperen.cryptocurrencydemoapp.model.ticker
  * Created by Alperen Erdoğan on 30.05.2024.
  */
 data class Volumes(
-    var d: Long = 0,
-    var m: Long = 0,
-    var w: Long = 0
+    var d: Double = 0.0,
+    var m: Double = 0.0,
+    var w: Double = 0.0
 )
